New notebooks, files
I am adding several notebooks although I will be deleting some of them.
Many of the new notebooks are dedicating to making figures for the paper
but as I settle on final figures I will delete the notebooks I'm not
using. Many of the notebooks are making the same figures and are just
combining them in different panel organizations.